Split core tests into multiple files
